requirements.txt
Flask
requests
beautifulsoup4
pandas
matplotlib
openpyxl
pytesseract
Pillow